update: 	[*] fix max-int64 overflow under 32-bit arch
Former-commit-id: 0763c6ebd38ac0a8a12cb75401da18b9dc42e4a1
Former-commit-id: 3482eabd9d2dba1df82929fb7e0138d36257d715 [formerly cc0422482da8c5809c4b1fe53bee3fdd7db63c46]
Former-commit-id: 86531beeb09a4b05c50ef7da8a701b6104d82364

+++ plugins/inputs/mysqlmonitor/util.go
@@ -58,7 +58,7 @@ func ParseValue(value sql.RawBytes) (interface{}, error) {
 
 	if val, err := strconv.ParseInt(string(value), 10, 64); err != nil {
 		if err, ok := err.(*strconv.NumError); ok && err.Err == strconv.ErrRange {
-			return math.MaxInt64, nil
+			return int64(math.MaxInt64), nil
 		}
 	} else {
 		return val, nil
